bot/requirements.txt
flask
langchain
faiss-cpu
pypdf
google-generativeai